날짜 버그관련 코드 추가, food탭 리팩토링 중
