+++ frontend/src/helper/calculateKcal.js
@@ -1,0 +1,3 @@
+export const sumKcal = arr => {
+  arr.reduce((pre, cur) => pre + cur)
+}
