+++ frontend/src/pages/DiaryPage/DiaryFood/FoodAdd.js
@@ -2,10 +2,10 @@ import React, { Component } from 'react'
 import { clearSelect } from '../../../actions/diaryFood'
 import { Button, Icon } from 'semantic-ui-react'
 import * as Style from './StyledDiaryFood'
-import DiaryFoodSearch from './DiaryFoodSearch'
-import DiaryFoodSearchModal from './DiaryFoodSearchModal'
+import FoodSearch from './FoodSearch'
+import FoodSearchModal from './FoodSearchModal'
 
-class DiaryFoodAdd extends Component {
+class FoodAdd extends Component {
   constructor(props) {
     super(props)
     this.state = {
@@ -22,9 +22,7 @@ class DiaryFoodAdd extends Component {
       <div>
         {this.state.isSearchMode ? (
           <div>
-            <DiaryFoodSearch
-              type={this.props.type}
-            />
+            <FoodSearch type={this.props.type} />
           </div>
         ) : (
           <div style={{ display: 'flex' }}>
@@ -40,7 +38,7 @@ class DiaryFoodAdd extends Component {
               />
               음식추가
             </Button>
-            <DiaryFoodSearchModal
+            <FoodSearchModal
               isSearchMode={this.toggleSearchMode}
             />
           </div>
@@ -50,4 +48,4 @@ class DiaryFoodAdd extends Component {
   }
 }
 
-export default DiaryFoodAdd
+export default FoodAdd
