+++ frontend/src/pages/DiaryPage/DiaryFood/FoodSearchDetails.js
